fix: use committed license stub for local dev and type checks
Replace the gitignored-only licenses.json import with a committed empty
package-licenses.json stub so TypeScript and Metro work without running
bun licences first. Release builds use licences:bundle to generate the
full list and copy it over the stub before bundling.
