+++ .github/workflows/ci.yml
@@ -25,9 +25,6 @@ jobs:
       - name: Install dependencies
         run: bun install --frozen-lockfile
 
-      - name: Generate license list
-        run: bun run licences
-
       - name: Run TypeScript checks
         run: bun tsc --noEmit
