+++ .github/workflows/release.yml
@@ -153,7 +153,7 @@ jobs:
         run: bun install --frozen-lockfile --ignore-scripts
 
       - name: Update license list
-        run: bun run licences
+        run: bun run licences:bundle
 
       - name: Set up JDK 17
         uses: actions/setup-java@ad2b38190b15e4d6bdf0c97fb4fca8412226d287 # v5.3.0
