+++ AGENTS.md
@@ -53,7 +53,6 @@ Run project scripts through Bun. The most relevant ones:
 
 ```bash
 bun install                  # install dependencies (use --frozen-lockfile in CI)
-bun licences                 # generate src/data/licenses.json once on fresh clones (generated, gitignored)
 
 bun prebuild:ios             # generate ios/ from Expo config (required on fresh clone)
 bun dev                      # start Expo dev server (with EXPO_USE_FAST_RESOLVER=1)
@@ -83,7 +82,7 @@ bun build:android            # local EAS build for Android
 bun build:ios                # prebuild + xcodebuild archive for iOS
 bun ship:ios                 # build + export + altool upload to App Store Connect
 bun build:all                # iOS + Android sequentially
-bun licences                 # regenerate src/data/licenses.json (generated, gitignored; never commit)
+bun licences:bundle          # regenerate package licenses for release builds
 bun changelog                # regenerate CHANGELOG.md via git-cliff
 bun atlas                    # Expo Atlas bundle analysis
 ```
@@ -151,8 +150,10 @@ Path aliases (defined in `tsconfig.json`):
 Generated and binary files:
 
 - `src/__generated__/` is GraphQL output. Never edit it directly; run `bun codegen`.
-- `src/data/licenses.json` is generated by `bun licences`, gitignored, and must not be committed.
-  Run `bun licences` after a fresh clone before running `bun tsc --noEmit` locally.
+- `src/data/package-licenses.json` is a committed empty stub (`{}`) used for local dev and
+  type-checking. Release builds run `bun licences:bundle`, which generates
+  `src/data/licenses.json` (gitignored) and copies it over the stub before bundling.
+  Import package licenses through `@/data/licenses`, not the JSON files directly.
 - `CHANGELOG.md` is generated by `bun changelog`.
 - Image, font, and SVG assets are tracked through Git LFS (`.gitattributes`). Make sure
   LFS assets are present before assuming an asset is missing or corrupt.
@@ -517,7 +518,9 @@ Android uses Material Symbols (custom font), Web uses `lucide-react-native`.
   compatibility. Run `bun pkgs` (which calls `expo install --check`) when in doubt.
 - **Don't edit generated files by hand.** Use `bun codegen` for `src/__generated__/`,
   `bun licences` for `src/data/licenses.json`, and `bun changelog` for `CHANGELOG.md`.
-- **Don't commit `src/data/licenses.json`.** It is generated on demand in CI/build pipelines.
+- **Don't commit `src/data/licenses.json`.** It is an intermediate build artifact.
+  `src/data/package-licenses.json` is committed as an empty stub; release pipelines
+  overwrite it ephemerally via `bun licences:bundle`.
 - **Don't commit the `ios/` folder** (only the `ios/ci_scripts` symlink for Xcode Cloud). It is
   gitignored; run `bun prebuild:ios` and change `app.config.json`, config plugins, or
   `config/ios-artifacts/` instead. Edit `config/ios-artifacts/ci_scripts/ci_post_clone.sh` for
