+++ package.json
@@ -12,15 +12,16 @@
 		"android": "expo run:android --device",
 		"ios": "expo run:ios --device",
 		"prebuild:ios": "git lfs pull && bunx expo prebuild -p ios",
-		"build:android": "bun licences && eas build --platform android --local",
+		"build:android": "bun licences:bundle && eas build --platform android --local",
 		"build:ios": "bun run prebuild:ios && xcodebuild archive -workspace ios/NeulandNext.xcworkspace -scheme NeulandNext -configuration Release -archivePath ios/build/NeulandNext.xcarchive",
 		"export:ios": "xcodebuild -exportArchive -archivePath ios/build/NeulandNext.xcarchive -exportPath ios/build/NeulandNext.ipa -exportOptionsPlist ios/exportOptions.plist && xcrun altool --upload-app -f ios/build/NeulandNext.ipa/NeulandNext.ipa -t ios -u $(security find-generic-password -s 'AppleAppStoreUpload' -g 2>&1 | grep -E 'acct' | awk -F= '{print $2}' | tr -d '\"') -p $(security find-generic-password -w -s 'AppleAppStoreUpload')",
-		"ship:ios": "bun licences && bun build:ios && bun export:ios",
-		"build:all": "bun licences && bun build:ios && bun build:android",
+		"ship:ios": "bun licences:bundle && bun build:ios && bun export:ios",
+		"build:all": "bun licences:bundle && bun build:ios && bun build:android",
 		"lint": "biome check",
 		"fmt": "biome check --fix",
 		"fmt:unsafe": "biome check --fix --unsafe",
 		"licences": "npm-license-crawler -onlyDirectDependencies -json src/data/licenses.json",
+		"licences:bundle": "bun run licences && cp src/data/licenses.json src/data/package-licenses.json",
 		"codegen": "graphql-codegen --config config/codegen.yml",
 		"changelog": "git cliff --config config/cliff.toml --output CHANGELOG.md",
 		"pkgs": "bunx expo install --check",
