+++ src/data/licenses.ts
@@ -1,0 +1,9 @@
+import type { PackageLicenses } from '@/types/licenses'
+
+import packageLicenses from './package-licenses.json'
+
+export type { LicenseEntry, PackageLicenses } from '@/types/licenses'
+
+const licenses = packageLicenses as PackageLicenses
+
+export default licenses
