+++ src/data/package-licenses.json
@@ -1,0 +1,1 @@
+{}
