+++ src/types/licenses.ts
@@ -1,0 +1,8 @@
+export interface LicenseEntry {
+	licenses: string
+	repository?: string
+	licenseUrl?: string
+	parents: string
+}
+
+export type PackageLicenses = Record<string, LicenseEntry>
